Update Android protocol implementation to the binary protocol

+++ android-client/app/src/main/java/com/klkblake/mm/SessionListener.java
@@ -4,10 +4,10 @@ package com.klkblake.mm;
  * Created by kyle on 17/10/15.
  */
 public interface SessionListener {
-    void receivedMessage(int id, long timestamp, boolean author, String message);
-    void receivedMessage(int id, long timestamp, boolean author, int numPhotos);
+    void receivedMessage(long id, long timestamp, boolean author, String message);
+    void receivedMessage(long id, long timestamp, boolean author, int numPhotos);
 
-    void receivedPart(int messageID, int partID);
+    void receivedPart(long messageID, int partID);
 
     void networkFailed(Throwable cause);
     void protocolViolation(String message);
